major restucturing of main project folder
