+++ solar_car_telemetry/src/inittemp.py
